+++ src/ClassicUO.Client/Common/Enums/MacroType.cs
@@ -110,5 +110,6 @@ public enum MacroType
     ToggleAutoLoot,
     ToggleVoiceRecognition,
     SetLastTarget,
-    ToggleAutoWalk
+    ToggleAutoWalk,
+    ToggleBandageAgent,
 }